Fix Ch9 CI - missing data-out dir
Note that a file must exist in data-out for it to be tracked by git.
The traditional strategy of adding a .gitignore file in the empty
directory is incompatible with the naive test reset strategy of
removing the entire directory. So next best option is to just populate
it with one of the expected output files.